Merged PR 2977: Remove partiotion key fields from parameters
Right, this is what I *think* is needed from testing with Dan but not 100% certain.

As far as I can tell, if a field is a partition key, it works if it's inline in the WHERE clause but not if it's parameterised - you just get an empty result.

This changes the relevant queries to fix this issue.

Related work items: #13962