Merged PR 11550: Fixed so that if a subsequent refresh is performed when a successor has been...
Fixed so that if a subsequent refresh is performed when a successor has been created which doesn't exist then the newly created successor is not deleted.
Fixed so that if a successor exists before refresh then it is persisted if the closure with successor variation is run.

Related work items: #63664, #63665