Merged PR 18009: Profiling - process dist periods with zero value if prior released version has value
Ensures that distributions periods for funding line with zero value are processed (and reprofiled) if the previous released version of the funding line has a value. This is required to ensure that any pre variation pointer information is maintained, but also used to send zero values to downstream systems to clear any previously sent values.

Related work items: #101798