Merged PR 18277: Profiling handle variation pointer being set before start of pattern
Don't add error if reprofiling not run due to pointer being before start of distribution periods. This can happen if a funding line has multiple profiles (for different provider types) with different start dates, for example local authority providers having periods starting in August, but academies not starting until September. If the variation pointer is set to September then it is valid for the profiling to have not run.

Related work items: #102518