Merged PR 19590: Clone Tool - additional calc with template history fix
Changed clone tool so that additional calculations which were previously template calculations in the source specification retain that history when they are created in source cloned specification. This is required as these types of calculations continue to use the funding stream namespace rather than the Calculations namespace used by normal additional calcs.

Also changed so that any calculations that throw an exception when being cloned are copied into the source specification commented out.

Related work items: #113026, #113027