Merged PR 20162: RM - Determining missing providers from groups fix
Fix issues that occur when provider moves from one funding group to another (for example if an academy changes trust).

First issue resolved; if a release contained changes to both the old and new group then the provider would silently fail to release.

Second issue resolved; if releases were being done in batches and a provider moved out of a grouping (A) but there were no changes to group A in the release then an old version of the provider would still be in group A (as expected), however when the group was updated in a subsequent release if the provider was not selected in the release then the provider would be incorrectly added to the new version of the group.

Related work items: #118099